requirements.txt
pinecone
time
openai
python-dotenv
langchain